Refactoring by moving User entity into separate file
